# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PaperSearch API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  db?: string;
  redis?: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface RateLimitError {
  error: string;
  retryAfter: number;
}

export interface PaperResult {
  id: number;
  paperId: string;
  title: string;
  authors: string;
  year?: number | null;
  abstract?: string | null;
  venue?: string | null;
  citationCount?: number | null;
  fieldOfStudy?: string | null;
  score: number;
}

export interface SearchResponse {
  query: string;
  total: number;
  page: number;
  pageSize: number;
  results: PaperResult[];
  suggestion?: string | null;
  latencyMs: number;
}

export interface SuggestResponse {
  suggestions: string[];
}

export type PaperDetailExternalIds = { [key: string]: unknown } | null;

export interface PaperDetail {
  id: number;
  paperId: string;
  title: string;
  authors: string;
  year?: number | null;
  abstract?: string | null;
  venue?: string | null;
  citationCount?: number | null;
  fieldOfStudy?: string | null;
  externalIds?: PaperDetailExternalIds;
  url?: string | null;
  indexedAt?: string | null;
}

export interface SimilarResponse {
  papers: PaperResult[];
}

export interface StatsResponse {
  total_docs: number;
  total_terms: number;
  avgdl: number;
  avgdl_title?: number;
  avgdl_abstract?: number;
  last_indexed_at?: string | null;
  totalDocs: number;
  totalTerms: number;
  avgdlTitle: number;
  avgdlAbstract: number;
  lastIndexedAt?: string | null;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 16 */
  apiKey: string;
}

export interface LoginRequest {
  email: string;
  apiKey: string;
}

export interface AuthResponse {
  token: string;
  userId: number;
  email: string;
  plan: string;
}

export type QueryLogEntryFilters = { [key: string]: unknown } | null;

export interface QueryLogEntry {
  id: number;
  userId?: number | null;
  query: string;
  filters?: QueryLogEntryFilters;
  resultCount?: number | null;
  latencyMs?: number | null;
  createdAt?: string | null;
}

export interface QueryLogsResponse {
  total: number;
  page: number;
  pageSize: number;
  logs: QueryLogEntry[];
}

export interface ReindexResponse {
  message: string;
  queued: number;
}

export type SearchPapersParams = {
  /**
   * Query string (supports -exclusions, "phrases", AND, field:prefix)
   * @minLength 1
   */
  q: string;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
  yearFrom?: number;
  yearTo?: number;
  fieldOfStudy?: string;
  /**
   * @minimum 0
   */
  minCitations?: number;
};

export type SuggestTermsParams = {
  /**
   * Prefix to complete
   * @minLength 1
   */
  q: string;
};

export type GetSimilarPapersParams = {
  /**
   * @minimum 1
   * @maximum 20
   */
  limit?: number;
};

export type GetQueryLogsParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
};