# lib/api-zod/src/generated/types/queryLogEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PaperSearch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueryLogEntryFilters } from "./queryLogEntryFilters";

export interface QueryLogEntry {
  id: number;
  userId?: number | null;
  query: string;
  filters?: QueryLogEntryFilters;
  resultCount?: number | null;
  latencyMs?: number | null;
  createdAt?: string | null;
}