# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PaperSearch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsResponse {
  total_docs: number;
  total_terms: number;
  avgdl: number;
  avgdl_title?: number;
  avgdl_abstract?: number;
  last_indexed_at?: string | null;
  totalDocs: number;
  totalTerms: number;
  avgdlTitle: number;
  avgdlAbstract: number;
  lastIndexedAt?: string | null;
}